imp: remove unused code snippets
