+++ .changelog/unreleased/improvements/869-codebase-cleanup.md
@@ -1,0 +1,2 @@
+- Remove unused code snippets and move serializer roundtrip test to `serializers.rs`
+  ([#869](https://github.com/cosmos/ibc-rs/issues/869))
